Provide optional search functionality in CarPlay
